requirements.txt
flask>=3.0
posthog>=3.0
python-dotenv>=1.0
